+++ Controller/KlarnaViewConfig.php
@@ -92,7 +92,9 @@ class KlarnaViewConfig extends KlarnaViewConfig_parent
 
             $url  = sprintf(KlarnaConsts::getFooterImgUrls(KlarnaUtils::getShopConfVar('sKlarnaFooterValue')), $sLocale);
             $from = '/' . preg_quote('-', '/') . '/';
-            $url  = preg_replace($from, '_', $url, 1);
+            if(KlarnaUtils::getShopConfVar('sKlarnaFooterValue') != 'logoFooter') {
+                $url  = preg_replace($from, '_', $url, 1);
+            }
 
             $response = array(
                 'url'   => $url,
@@ -104,33 +106,7 @@ class KlarnaViewConfig extends KlarnaViewConfig_parent
             $response['script'] = KlarnaUtils::getShopConfVar('sKlarnaMessagingScript');
         }
 
-        if(KlarnaUtils::getShopConfVar('sKlarnaFooterPromotion')) {
-            $response['promotion'] = KlarnaUtils::getShopConfVar('sKlarnaFooterPromotion');
-        }
-
-
         return $response;
-    }
-
-    /**
-     *
-     */
-    public function getKlarnaHomepageBanner()
-    {
-        if (KlarnaUtils::getShopConfVar('blKlarnaDisplayBanner')) {
-            $oLang = Registry::getLang();
-            $lang  = $oLang->getLanguageAbbr();
-
-            $varName = 'sKlarnaBannerSrc' . '_' . strtoupper($lang);
-            if (!$sBannerScript = KlarnaUtils::getShopConfVar($varName)) {
-                $aDefaults     = KlarnaConsts::getDefaultBannerSrc();
-                $sBannerScript = $aDefaults[$lang];
-            }
-
-            return str_replace('{{merchantid}}', KlarnaUtils::getShopConfVar('sKlarnaMerchantId'), $sBannerScript);
-        }
-
-        return false;
     }
 
     public function getOnSitePromotionInfo($key, $detailProduct = null)