+++ Tests/Unit/Controller/KlarnaViewConfigTest.php
@@ -97,14 +97,6 @@ class KlarnaViewConfigTest extends ModuleUnitTestCase
         $this->assertFalse( $oViewConfig->getIsAustria());
     }
 
-    public function getKlarnaHomepageBannerDataProvider()
-    {
-        return [
-            [true, 'mid'],
-            [false, 'mid2']
-        ];
-    }
-
     public function isDEDataProvider()
     {
         return [
@@ -133,28 +125,6 @@ class KlarnaViewConfigTest extends ModuleUnitTestCase
         $oViewConfig = $this->getMockBuilder(ViewConfig::class)->setMethods(['getUser'])->getMock();
         $oViewConfig->expects($this->once())->method('getUser')->willReturn(null);
         $this->assertTrue( $oViewConfig->getIsGermany());
-    }
-
-    /**
-     * @dataProvider getKlarnaHomepageBannerDataProvider
-     * @param $displayBanner
-     * @param $merchantId
-     */
-    public function testGetKlarnaHomepageBanner($displayBanner, $merchantId)
-    {
-
-        $this->getConfig()->saveShopConfVar('bool', 'blKlarnaDisplayBanner', $displayBanner, $this->getShopId(), 'module:tcklarna');
-        $this->getConfig()->saveShopConfVar('str', 'sKlarnaMerchantId', $merchantId, $this->getShopId(), 'module:tcklarna');
-
-        $oViewConfig = oxNew(ViewConfig::class);
-        $result = $oViewConfig->getKlarnaHomepageBanner();
-
-        if($displayBanner){
-            $this->assertContains($merchantId, $result);
-        } else {
-            $this->assertEquals(null, $result);
-        }
-
     }
 
     public function showCheckoutTermsDataProvider()
@@ -257,14 +227,10 @@ class KlarnaViewConfigTest extends ModuleUnitTestCase
     {
         return [
             ['KP', 0, 'longBlack',false,false, false],
-            ['KP', 1, 'logoBlack',false,false, false],
-            ['KP', 2, 'logoBlack',false,false, [
-                'url' => '//cdn.klarna.com/1.0/shared/image/generic/logo/en_gb/basic/logo_black.png',
-                'class' => 'logoBlack'
-            ]],
-            ['KP', 2, 'logoWhite',false,false, [
-                'url' => '//cdn.klarna.com/1.0/shared/image/generic/logo/en_gb/basic/logo_white.png',
-                'class' => 'logoWhite'
+            ['KP', 1, 'logoFooter',false,false, false],
+            ['KP', 2, 'logoFooter',false,false, [
+                'url' => 'https://x.klarnacdn.net/payment-method/assets/badges/generic/klarna.svg',
+                'class' => 'logoFooter'
             ]],
             ['KCO', 0, 'longBlack', false,false, false],
             ['KCO', 1, 'longBlack',false,false, [
@@ -283,19 +249,14 @@ class KlarnaViewConfigTest extends ModuleUnitTestCase
                 'url' => '//cdn.klarna.com/1.0/shared/image/generic/badge/de_de/checkout/short-white.png?width=312',
                 'class' => 'shortWhite'
             ]],
-            ['KCO', 2, 'logoBlack',false,false, [
-                'url' => '//cdn.klarna.com/1.0/shared/image/generic/logo/en_gb/basic/logo_black.png',
-                'class' => 'logoBlack'
+            ['KCO', 2, 'logoFooter',false,false, [
+                'url' => 'https://x.klarnacdn.net/payment-method/assets/badges/generic/klarna.svg',
+                'class' => 'logoFooter'
             ]],
-            ['KCO', 2, 'logoWhite',false,false, [
-                'url' => '//cdn.klarna.com/1.0/shared/image/generic/logo/en_gb/basic/logo_white.png',
-                'class' => 'logoWhite'
-            ]],
-            ['KCO', 2, 'logoWhite','script','promo', [
-                'url' => '//cdn.klarna.com/1.0/shared/image/generic/logo/en_gb/basic/logo_white.png',
-                'class' => 'logoWhite',
-                'script' => 'script',
-                'promotion' => 'promo'
+            ['KCO', 2, 'logoFooter','script','promo', [
+                'url' => 'https://x.klarnacdn.net/payment-method/assets/badges/generic/klarna.svg',
+                'class' => 'logoFooter',
+                'script' => 'script'
             ]],
         ];
     }
@@ -324,7 +285,6 @@ class KlarnaViewConfigTest extends ModuleUnitTestCase
         $this->getConfig()->saveShopConfVar('str', 'sKlarnaFooterValue', $klFooterValue, $this->getShopId(), 'module:tcklarna');
 
         $this->getConfig()->saveShopConfVar('str', 'sKlarnaMessagingScript', $klScript, $this->getShopId(), 'module:tcklarna');
-        $this->getConfig()->saveShopConfVar('str', 'sKlarnaFooterPromotion', $klPromo, $this->getShopId(), 'module:tcklarna');
         $this->setModuleMode($mode);
 
         $oViewConfig = oxNew(ViewConfig::class);